Zero unsafe issue fixed

+++ src/tauargus/service/TableService.java
@@ -285,7 +285,7 @@ public class TableService {
             if (frequencyMarge[0] < frequencyMarge[1] && table.holding && table.minFreq[1] > 0) {
                 frequencyMarge[0] = frequencyMarge[1];
             }
-            if (!tauArgus.SetTableSafety(i, table.domRule, table.domN, table.domK, table.pqRule, table.pqP, table.pqQ, table.pqN, minFreq, table.piepPercentage, table.piepMarge, table.piepMinFreq, isPiep, table.weighted, table.weighted, table.holding, table.zeroUnsafe, false, 10, (int) table.zeroRange, table.manualMarge, frequencyMarge)) {
+            if (!tauArgus.SetTableSafety(i, table.domRule, table.domN, table.domK, table.pqRule, table.pqP, table.pqQ, table.pqN, minFreq, table.piepPercentage, table.piepMarge, table.piepMinFreq, isPiep, table.weighted, table.weighted, table.holding, table.zeroUnsafe, false, 10, table.zeroRange, table.manualMarge, frequencyMarge)) {
                 throw new ArgusException("SetTableSafety went wrong for table" + (i + 1));
             }
         }
